Move Mock Provider to examples/provider
The Mock Provider needs to be able to be used by tests. By keeping it in
an internal/ directory, it means we cannot use it (easily) for tests.
