+++ cmd/virtual-kubelet/register.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/virtual-kubelet/virtual-kubelet/cmd/virtual-kubelet/internal/provider"
-	"github.com/virtual-kubelet/virtual-kubelet/cmd/virtual-kubelet/internal/provider/mock"
+	"github.com/virtual-kubelet/virtual-kubelet/examples/providers/mock"
 )
 
 func registerMock(s *provider.Store) {
