+++ examples/providers/mock/mock.go
